feat(Lotto): 중복된 로또 번호가 있을경우 에러가 발생하는 기능 구현
