+++ README.md
@@ -20,8 +20,8 @@
    - [ ] ⚠️ 1000원 단위로 나누어 떨어지지 않을 경우 에러가 발생한다.
  - [ ] 구매 횟수만큼 로또를 발행하고 오름차순으로 정렬한다.
  - [ ] 입력받은 당첨 번호를 검증한다.
-   - [ ] ⚠️ 중복된 번호가 있을경우 에러가 발생한다.
-   - [ ] ⚠️ 6개가 안될경우 에러가 발생한다.
+   - [x] ⚠️ 중복된 번호가 있을경우 에러가 발생한다.
+   - [x] ⚠️ 6개가 안될경우 에러가 발생한다.
    - [ ] ⚠️ 1 ~ 45의 범위를 넘어갈경우 에러가 발생한다.
  - [ ] 입력받은 보너스 번호를 검증한다.
    - [ ] ⚠️ 입력받은 당첨번호와 중복될경우 에러가 발생한다.
